Authorization using Gate

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -13,6 +13,11 @@ class AttendeeController extends Controller
 {
     use CanLoadRelationships;
     private array $relations = ['user'];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -56,8 +61,10 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete-event', [$event, $attendee]);
+
         $attendee->delete();
 
         return response(status: 204);
